requirements.txt
typer==0.9.0
rich==13.3.5
PyYAML==6.0